Incompatibilities between BSD sed and GNU sed
Whereas GNU sed's -i/--in-place option optionally takes an argument
(which is a suffix, and means to modify the input file in place if
you omit an argument), BSD sed's -i option always takes an argument
even if it's an empty string (e.g., -i '').

My solution is install GNU sed using Homebrew on macOS build.

Note that omitting --with-default-names will prepend "g" to the
command, i.e., gsed.

[changelog skip]